BIGGGGG CHANGES PLS READ DESC
1. Tentacle Appendage.

Refinement: Remove Physics.Overlapcircle and replace it with raycast.

Why:
Because player can grab onto the object by clicking the surrounding of the object. we want the player to click on the object itself to move it.

Mass Segment:

Created: Breakable Object with HPbar visual tagging along with it.
Why:
To see visually if the projectile mechanic is working correctly.

To do:
HP bar visual to be refine visually. The rectangle seems to be abit deformed.

Mass Segment.
Created mass segment that can either boost/deduct stats from the players. Based on the debug log, seems like its working.

Todo:
Gonna test it visually see if theres feels to the changes.

Mass Segment's Script explanation:
1. Mass Stats (To hold & calculate the stats changes)
2. Mass Segments (To manage the stats)
3. Pickup Detector (To detect and pickup the segments0
4. Player Stats (To scales according to the player stats)