Failed Subject Enemy
A failed creature experiment that was tortured and left behind in the lab. It moves unpredictably and attacks with erratic charges. Players would need to move interactable objects to dodge, strike, or manipulate the environment against them.

Foundation Done. (Chase/Patrol)
1. Erratic charges (No)

Additional, since it can charge, expect bounce off force to the player.