Update Enemy (FailedSubjects)
A failed creature experiment that was tortured and left behind in the lab. It moves unpredictably and attacks with erratic charges. Players would need to move interactable objects to dodge, strike, or manipulate the environment against them.

Player can now move grab object to hide it from the enemy

Enemy patrol Point A to Point B.
if obj infront of the enmy, it returns to its respective points
if no obj, it moves to point b ( patrolling)
if player found, erratic charge toward players

if players comes in contact with enemy, it got knockback due to the charging.