Update Drone
A series of small agile robotics helpers designed by the Scientist to capture Gloo. It patrols corridors and attempts to push Gloo and its segment. Players can absorb and reabsorb segments to avoid capture or use momentum and mass to knock drones into walls or each other.

Player can now move grab object to hide it from the enemy

Enemy patrol Point A to Point B.
if obj infront of the enmy, it returns to its respective points
if no obj, it moves to point b ( patrolling)
if player found, enemy chase him.